requirements.txt
blinker==1.6.3
chess==1.10.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-chess==1.999
python-shogi==1.1.0
Werkzeug==3.0.0
